fix(reconcile-pagination): extract cursor helpers to lib/d1/reconcile-cursor.ts
Next.js App Router rejects non-Route exports from route.ts files at build
time ("'encodeCursor' is not a valid Route export field"). Moves
InboxCursorState, InboxPartialCounts, encodeCursor, decodeCursor to a
new lib/d1/reconcile-cursor.ts module.

This fixes the Cloudflare Workers Builds CI failure that arc0btc flagged
as the remaining blocker on PR #701. Build verified locally.

Also tightens decodeCursor structural validation to require txidCounts
field per arc0btc cycle-2 review suggestion.

Also fixes a TS narrowing issue on KVNamespaceListResult discriminated
union (page.cursor only exists on the list_complete=false variant).

Refs #684.
